requirements.txt
# ===========================================
# Wakili Legal Chatbot - Requirements
# ===========================================

# Core
python-dotenv>=1.0.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anyio>=4.0.0

# LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langchain-classic>=0.0.1

# Vector Store
langchain-chroma>=0.1.0
chromadb>=0.4.0

# Embeddings & Reranker
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0

# LLM Provider (Groq)
langchain-groq>=0.1.0

# BM25 Keyword Search
rank-bm25>=0.2.2

# Numerical
numpy>=1.24.0

# HTTP
requests>=2.31.0
